refactor: main.py의 비즈니스 로직 및 DB 처리 코드 모듈을 분리한다
